chore: move autorestake button to staking page
